Improve status command
If there are too many routes with performance issues to list individually, the number of routes with issues in each category will be displayed instead, and if more than 30% of all routes are experiencing issues, the bot will just respond with an increasing amount of 🔥 emoji. Also resolved some cosmetic problems introduced in my last change to this command.